fix: Return promise in put
This should probably be covered by the test suite in the future,
however the problem was that when the writing of the changes takes time
micro-analytics will not wait for it to finish. In the tests for
micro-analytics were we mock out flat-file-db and add latency this
breaks several tests.